Release 0.5.1: align build_sites output with data_processing survivors
Follow-up refinement to 0.5.0. End-to-end validation on Bahamas
uncovered that only ~11 % of the 2618 mask-verified candidates
returned by build_sites actually reached Time_series_data_*.h5 (290
survived data_processing, matching the pre-0.5.0 count). Three
compounding issues, all fixed here:

1. Mask depth was hardcoded to 1062.44 m in build_sites. Correct for
   default runs, silently misaligned when a caller overrode
   SeawaterPipes.cw_inlet_length. regional.run_regional_analysis now
   forwards cmems_verify_depth=float(inputs['length_CW_inlet']) to
   load_sites, so the mask always tracks the depth CMEMS will be
   queried on.

2. build_sites placed candidates on multiples of 1/12° measured from
   the region bbox origin, but the CMEMS native grid has a global
   half-cell offset (~9 km at the equator). The exact (lon, lat)
   lookup in cmems._extract_year_data missed almost every candidate.
   filter_sites_by_cmems_mask now snaps each surviving site's
   coordinates to the actual CMEMS cell centre (rounded to 3 decimals
   to match the downstream rounding) and dedupes when multiple GEBCO
   candidates snap to the same cell.

3. cmems._extract_year_data rounded coordinates at float32 precision:
   np.round(np.float32(23.833), 3) returns 23.83300018310547 because
   23.833 has no exact float32 representation. The sites_dict lookup
   key downstream promotes that value to float64 preserving the
   ~1e-7 offset; the site-side keys, built from float64 coordinates,
   round cleanly to 23.833, so the two never matched. Coordinates
   are now upcast to float64 before rounding.

End-to-end regression on Bahamas at defaults, cold cache:
  build_sites:   2618 sites
  h5 downstream: 2618 sites  (was 290 in 0.5.0)
  match:         perfect (set difference: 0 extra, 0 missing)

Test C (mask depth propagation) unchanged and deterministic across
runs: 3381@800m, 2618@1062m, 1328@2000m.

+++ otex/data/cmems.py
@@ -324,8 +324,19 @@ def _extract_year_data(year_files, sites_dict, time_origin):
 
     for f in year_files:
         ds = netCDF4.Dataset(f, 'r')
-        latitude = np.round(ds.variables['latitude'][:], 3)
-        longitude = np.round(ds.variables['longitude'][:], 3)
+        # Upcast float32 → float64 BEFORE rounding: CMEMS stores the
+        # coordinate axes as float32, and ``np.round(<float32>, 3)``
+        # returns the nearest float32 to X.XXX — which for values like
+        # 23.833 is not exactly representable, and comes back as
+        # 23.83300018310547. The (lon, lat) lookup key downstream is
+        # built with ``float(lon_val)`` which promotes that same
+        # imprecise value to float64, so the ``sites_dict`` key
+        # (built from float64 site coords via ``np.round(..., 3)``,
+        # which does round cleanly at float64 precision) can't match
+        # it. Promoting first makes the round operate at float64
+        # precision on both sides and the ~1e-7 offset disappears.
+        latitude = np.round(ds.variables['latitude'][:].astype(np.float64), 3)
+        longitude = np.round(ds.variables['longitude'][:].astype(np.float64), 3)
         depth = int(ds.variables['depth'][:])
         T_water = np.asarray(ds.variables['thetao'][:], dtype=np.float64)
         # Collapse the depth axis (typically length 1) and the spatial
